build: allow passing nodejs interpreter command line for scripts
Distros need to change the command of nodejs interpreter, as it
conflicts with other packages. As a consequence, executable scripts
need to be changed, so the correct interpreter is called by the kernel.

Instead of leaving the distros alone with the burden of patching behind
the upstream forever, solving the problem on npm installation, while
keeping compatiblity w/ legacy installations:

* the environment variable $NODEJS is taken for the nodeJS command
  name (w/o pathname) - default is 'node'.
* the environment variable $NODEJS_INTERPRETER is taken for the
  actual command line in the script heads - default is
  '/usr/bin/env '+<$NODEJS>

A simple usecase could be just setting NODEJS to 'nodejs', then
NODEJS_INTERPRETER would default to '/usr/bin/env nodejs', which
already should catch most usecases. In case of distro needs to
specify a different command line (eg. full pathname or extra
parameters), it can define NODEJS_INTERPRETER.

2DO: add an post-processing logic that also fixes up 3rdparty
scripts that are installed via NPM.